backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
defusedxml==0.7.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
fonttools==4.58.1
idna==3.10
iniconfig==2.1.0
joblib==1.5.1
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.6
oauthlib==3.2.2
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pluggy==1.6.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.2.3
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
stripe==12.2.0
threadpoolctl==3.6.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
nltk
rake-nltk
